Fix code snippets in Fivetran integration docs (#27831)
## Summary & Motivation
Add missing / implied `Definition` to an example code snippet for adding column-level metadata to a Fivetran asset. This makes it so users can copy-paste the code snippet into their Dagster project and use it without needing to add the `Definition` themselves.

Also adds some comments to make it clear where a user needs to replace placeholder with real value for Fivetran connector id.

## How I Tested These Changes
- Set up a Fivetran connector
- Set up new Dagster project
- Copy the example into `definitions.py`, replacing values as specified.
- Run `dagster dev`, verify no errors
- Verify the Fivetran asset has column-level metadata

## Changelog

> Insert changelog entry or delete this section.