Add a validation step to DA that raises an exception and retries if asset keys were produced
Summary:
This doesn't capture every possible race condition that can happen in terms of assets being removed from the graph in the middle of a DA tick, but captures one common failure scenario, where a long DA tick produces assets that are no longer in the graph, the cursor is written, the no longer valid run requests are retried a few times, and then the DA logic moves on, potentially skipping work.

Fixed an issue where removing an asset in the middle of a Declarative Automation tick could sometimes cause some evaluations to be skipped.