Handle source code scaling (#27419)
## Summary & Motivation
By default, when over a certain number of DAGs, switch to not retrieving
source code. I also make this default state overridable for users who
determine that the tradeoff is worth it.

Similar to the downstack, this significantly speeds up initial load, for
200 dags from 50s down to 20-odd seconds.

## How I Tested These Changes
Additional tests changing the default scaling limits and ensuring that
they are respected.