Switch dagster dev IPC mechanism to pipe from signal (#27750)
## Summary & Motivation

`dagster dev` spins up webserver and daemon subprocesses. Currently, on
being interrupted, it attempts to shutdown those processes by sending an
interrupt signal to each of them.

This is an unreliable mechanism for shutting down subprocesses on
Windows, because the ability of one process to signal another process is
dependent on them sharing a console. This is a particularly unreliable
condition in a CI environment. Quoting ChatGPT:

>On Windows, CTRL_BREAK_EVENT (and similarly CTRL_C_EVENT) only works
when the calling process and the target processes share a console. By
default, Python’s subprocess.Popen with CREATE_NEW_PROCESS_GROUP spawns
the child into its own new process group but does not automatically
attach it to the same console as the parent (and in many CI
environments, there may be no real console at all). As a result, sending
CTRL_BREAK_EVENT to that child process group essentially does
nothing—Windows never delivers the interrupt to those child processes,
and they keep running.

There are also other weird effects that arise in niche conditions from
sending signals on windows. For instance, if a windows process gets
interrupted in the middle of a `print()` call, it disrupts stdout and
raises a confusing `OSError: invalid argument`.

While there is probably some way to robustly handle signals on windows,
I have encountered enough puzzling behavior that it seemed worthwhile to
explore an alternative and more reliable approach to IPC.

This PR adds a new "shutdown pipe" API useful for IPC. A process can
monitor a pipe for a shutdown signal, which can be configured to trigger
an interrupt to the _current_ process. Importantly, on windows we can
entirely skip signals by making use of `_thread.interrupt_main()`, which
will raise a `KeyboardInterrupt` in the main thread the same way that
receiving `SIGINT` etc do, but without any of the other signal side
effects.

`dagster dev`, `dagster-webserver` and `dagster daemon` have all had a
(hidden) option added called `--shutdown-pipe`. This accepts a readable
file descriptor which can be passed into the shutdown pipe API. The idea
is that a parent process can invoke the child and pass
`--shutdown-pipe`, then send the shutdown message over the pipe in lieu
of relying on the OS to deliver an inter-process signal. `dagster dev`
has been changed to use this mechanism for its child `dagster-webserver`
and `dagster daemon` processes.

## How I Tested These Changes

New tests + existing dagster dev tests (on windows also)