Fix asset backfill subset logging
Summary:
AssetGraphSubset.empty() returns an empty subset. What we wanted here was a boolean that tells you whether the subset was empty. This was (in master) causing asset backfills to always log that there was nothing to request, even if they then went on to request things

Test Plan:
New test case, Launch an asset backfill and view the logs, verify that output now includes the asset partitions being requested.