[Asset Graph] Remove throttling to fix double loading spinner (#27890)
## Summary & Motivation

Recently I added throttling to the computeGrahData webworker call
because we were recalculating the graph automatically on key-press. We
removed that behavior and went back to requiring "Enter" to submit so
I'm removing the throttling here which was causing a double loading
spinner in some situations.

Also removed the fallback behavior where we interpret the asset
selection syntax using the old parser.

## How I Tested These Changes

jest + manual testing