[ui] Upgrade Storybook (#27808)
## Summary & Motivation

Update Storybook in ui-core and ui-components.

`@storybook/core` has a dependency that needs to be updated according to dependabot, but it doesn't look like a simple resolution override, so we're waiting on Storybook to release a version with the change. In the meantime, I had already run the upgrade scripts to see if it would pick up the new dep, and the new version seems a little faster than it had been.

## How I Tested These Changes

`yarn storybook` on both packages.