[SVGViewport] Add provider to allow preserving state when component is un-mounted and re-mounted (#27893)
## Summary & Motivation

Currently in the asset-graph we unmount the SVGViewport component to
show a loading spinner whenever the graph changes in anyway (eg. due to
filters changing or group expansion/collapse). This causes us to lose
the X/Y coordinate and Zoom. To fix that I'm adding an OPTIONAL provider
component for SVGViewport. By default SVGViewport wraps itself with this
provider, however the provider will use the state of any parent provider
if there is one. This allows the asset graph to remove the SVGViewport
component and add it back while preserving its state.

## How I Tested These Changes

Tested with group expansion/collapse:


## Changelog
[ui] Expanding/collapsing groups in the Asset Graph will no longer reset
your zoom.