[components] Better error message for failing entry points (BUILD-647) (#27530)
## Summary & Motivation

Currently if there is an error when loading a `dagster-components` entry
point, the whole process terminates unceremoniously with an ugly error.

This PR changes this behavior:

Entry points are loaded in a try-catch block. We wrap any error when
loading in a `ComponentsEntryPointLoadError`. This and the backtrace
flow up to the output of the instigating `dg` command, which then caps
it off with a clean error message about the `dagster-components` command
failing.

This has the two keys properties of being debuggable and _ending_ the
output with a clear error message.

## How I Tested These Changes

New unit test and manually (running `dg component-type list` in an env
with a bad entry point).