Exclude failed subset when logging "Assets requested or in progress" in asset backfills (#27900)
## Summary & Motivation
We were including failed partitions in both the in progress list and the
failed list. Now we aren't.

Test Plan:
Launch a backfill that fails an asset, view backfill log output
New test case

## Changelog
Fixed an issue where asset backfills included failed partitions in the
in-progress list in logging output.