[Selection Input] Don't show auto-complete dropdown when programmatically updating the value (#27915)
On the run page we programmatically update the selection syntax when you
click on a step in the Gantt chart. Currently, when this happens we end
up making the auto-complete results dropdown popup. This is
unintentional behavior. Fix that behavior by not showing the dropdown if
we're programmatically updating the value.