[Run Timeline] Don't drop job data if locations within the FutureTicksQuery fails (#27944)
## Summary & Motivation

Currently, if the FutureTicksQuery itself fails we recover by returning
the ongoing runs and complete runs data that we have.

However, if the query doesn't fail then we end up iterating over all of
the locations within it and constructing a `jobs` array with all of the
rows of the timeline. The problem is that this construction relies on
the jobs returned by FutureTicksQuery. If FutureTicksQuery doesn't
return a particular job then we drop the data for that job completely.
To fix this track which keys we've added via the FutureTicksQuery and
then do a second pass where we add data for any jobs that were not in
the FutureTicksQuery

## How I Tested These Changes

Loaded the Run timeline for a customer with a failing location entry


## Changelog

> Insert changelog entry or delete this section.