[components] Rename params to attributes (#27958)
## Summary & Motivation

From our notion doc:

Author: @Nick Schrock 

Consulted: @Ben Pankow @Sean Mackesey 

The top-level key of in `component.yaml` used to be `params:` which
stems from our naming convention of the pydantic classes. They were
suffixed with `Params`; now they are suffixed by `Schema`. Somehow
`params` did not feel right as they usually apply to a function not an
object.

Options considered

- `params`, as stated above, sounds like the parameters to a function.
- Properties in Python is a readonly function. (`props` or `properties`)
- `args` although technically is Pythonic and accurate, the difference
between params and args is only enforced by the most exacting of us
(e.g. the pyright guy)
- `fields` was considered by we think that applies to definitions, not
instantiations.
- `config` was floated but collides with existing notion of config
(which if could travel in time would name params, ironically)

We decided on `attributes` because it is a term of art in JSON and
Python that represents a key value pair. That means they are “values”
all the way down, which is true from the standpoint of a consumer of a
component

Before

```yaml
type: sling_replication_collection@dagster_components

params:
  replications:
    - path: ./replication.yaml
      asset_attributes:
        key: "foo/{{ stream_definition.config.meta.dagster.asset_key }}"
  sling:
    connections:
      - name: DUCKDB
        type: duckdb
        instance: <PLACEHOLDER>
        password: "{{ env('SOME_PASSWORD') }}"

```

After:

```yaml
type: sling_replication_collection@dagster_components

attributes:
  replications:
    - path: ./replication.yaml
      asset_attributes:
        key: "foo/{{ stream_definition.config.meta.dagster.asset_key }}"
  sling:
    connections:
      - name: DUCKDB
        type: duckdb
        instance: <PLACEHOLDER>
        password: "{{ env('SOME_PASSWORD') }}"

```

## How I Tested These Changes

BK

---------

Co-authored-by: benpankow <ben@dagsterlabs.com>