[components] Remove `extend` key from config (BUILD-635) (#27998)
## Summary & Motivation

Now that we converged on the concept of workspaces, we no longer need
the `extend` key in config for inheritance. This PR simplifies config
merging:

- If the closest config file to CWD is a workspace config file, that's
our only config
- If the closest config file to CWD is not a workspace config, then look
for the workspace config and merge the two files if you find one

Also adds some previously missing tests for config inheritance and
invalid config keys.

## How I Tested These Changes

Unit tests.