Fix issue where seeding the random number generator during import or resource init time could cause every step to write to the same compute log key (#28017)
## Summary & Motivation
If you had a resource that seeded the python RNG (or RLG in this case)
in could cause every step to be assigned the same key, causing them to
overwrite each other. This fixes that by temporarily reseeding the RNG
during log key construction.

## How I Tested These Changes
New test case that was failing before

## Changelog
Fixed issue where seeding the random number generator during code import
or when initializing a resource could cause every step to write to the
same stdout or stderr key.