[components] Standardize duckdb table width when generating CLI snippet output (#27997)
## Summary & Motivation

Standardize output width for duckdb commands in docs snippets. We do
this by executing them in a pseudoterminal using `pexpect` and setting
duckdb's `.maxwidth` setting.

DuckDB's `.maxwidth` setting is extremely finicky and annoying. It is
often silently ignored. The only way I could get it to reliably actuate
was when running interactively from the command line. It seems to change
it's behavior based on the kind of output stream it's hooked up to.

The solution here is unfortunately complex, but so far as I can tell
there is no way at all to get `duckdb` to respect `.maxwidth` outside of
a terminal environment. I tried using an `-init` script and many other
things. It just silently ignores `.maxwidth` unless you are in a
terminal. So we simulate a terminal with `pexpect`, which works. This
same approach may prove useful in the future for other CLIs that try to
detect a terminal.

## How I Tested These Changes

Regenerated snippets locally, no changes-- previously it would generate
a longer duckdb table than would be generated in CI.