[dagster-powerbi] Handle invalid asset owners for reports and semantic models (#28027)
## Summary & Motivation

A bug was reported by a user, where the value of semantic model's
`configuredBy` was a UUID instead of a valid email. The `configuredBy`
field is used to set the asset owner value in AssetSpec - this causes an
exception because a UUID is not a valid asset owner.

> configuredBy | string | The dataset owner

Looking at the `configuredBy` field in PowerBI API docs, the string
format for this field is not mentioned. See in [this
page](https://learn.microsoft.com/en-us/rest/api/power-bi/admin/datasets-get-datasets-in-group-as-admin).

This is also true for report's `createdBy`.

To fix this, let's use `is_valid_asset_owner` for both cases to make
sure we don't pass an invalid asset owner value to AssetSpec, which
would cause an exception.

## How I Tested These Changes

Additional tests with BK