Skip sensor runs from same tick that use same run key (#27990)
## Summary & Motivation

Sensors are supposed to skip subsequent `RunRequests` that have the same
run key as a previous run. This works across ticks but not for
`RunRequests` within the same tick-- currently two runs will be launched
with the same key from within the same tick.

This PR amends the sensor daemon logic to ensure that sensor
`RunRequests` with the same run key are handled the same as they are
across runs.

User @Gw1p submitted an initial attempt
[here](https://github.com/dagster-io/dagster/pull/27322) at fixing this
but an alternative approach was required to ensure identical handling of
cross-tick and same-tick duplicates.

## How I Tested These Changes

New unit test.

## Changelog

Fixed a bug where a sensor emitting multiple `RunRequests` with the same
run key within a single tick would cause two runs with the same key to
be executed. Now only the first run will be executed.

Co-authored by: Anton <20753476+Gw1p@users.noreply.github.com>