[ui] Show warning on partitions that did not materialize due to canceled runs (#28045)
## Summary & Motivation

It is confusing for partitions that were not materialized due to canceled runs do not show any messaging indicating that a failure or cancellation occurred on the asset partition detail page.

This is because we only show the banner if the run failed. Update this to show a banner if it was canceled as well. Use a "warning" intent for the cancellation banner, to help visually distinguish it from the failed run.

Also updated the copy a bit, since it's the runs that fail or are canceled.

<img width="535" alt="Screenshot 2025-02-25 at 11 02 31" src="https://github.com/user-attachments/assets/a1605ebd-6e9a-4b96-bb16-a8cd96288b15" />
<img width="469" alt="Screenshot 2025-02-25 at 11 00 19" src="https://github.com/user-attachments/assets/09c4180f-c0e7-40fb-9826-43357c275f7b" />

## How I Tested These Changes

View partition that failed to materialize due to cancellation. Verify that the yellow banner appears.

View partition that failed to materialize due to failed run. Verify that the red banner appears.

## Changelog

[ui] For asset partitions that fail to materialize due to a run cancellation, show this information on the asset partition detail page.