[ui] Fix links on breadcrumb overflow menu (#28047)
## Summary & Motivation

Resolves https://github.com/dagster-io/dagster/issues/26471.

Repair links in the breadcrumb overflow menu.

These are currently unaware of react-router context, so they have no access to the `basePath` value. Unfortunately, Blueprint doesn't appear to give us any way to override how these menu items are rendered -- the `overflowProps` specifically omit `overflowRenderer` and `visibleItemRenderer` from being specified by client code.

To work around this, apply the `basePath` value to all of the hrefs to make it available to the overflow menu, and remove it from rendered breadcrumb elements, which use react-router Link components.

## How I Tested These Changes

View an asset with a long key path, shrink the viewport horizontally to make the breadcrumbs collapse. Open the overflow menu, verify that links are correctly prepended with the `basePath` value. Hover on breadcrumb items, verify that their hrefs are also correct, with `basePath` applied.

## Changelog

[ui] Fixed link path prefixes on breadcrumb overflow menu.