[components] Change top level attribute "transforms" to "post_processors" (BUILD-725) (#28039)
## Summary & Motivation

This changes the name of the top-level "transforms" attribute of the sling and dbt components to "post_processors". "transforms" was quite confusing because it communicates nothing about when it is applied relative to the dbt/sling translator, which is configured by `"asset_attributes"`. `"post_processors"` still requires some explanation (we should probably inject explanatory commentary when generating the component.yaml) but is more clear.

## How I Tested These Changes

Existing test suite.