[components] Add attributes to control metadata on materialized sling assets (#28041)
## Summary & Motivation

This adds a top level attribute `include_metadata` on `SlingReplicationCollection`. This is a list that can include `"column_metadata"` and `"row_count"` to match the methods on the `SlingEventIterator`. The default value includes both these fields.

Also cleans up the test code for sling a little bit.

## How I Tested These Changes

New unit test.