handle missing run group when determining if a run should be retried (#28053)
## Summary & Motivation
If the run group doesn't exist we were raising an error. This catches
the error and logs an engine event. In the case where the run group
doesn't exist, the run is not retried.

A run group wouldn't exist if the root run for the run group is deleted 

## How I Tested These Changes
unit tests

## Changelog

Fixed a bug where if a run was deleted while the re-execution system was determining whether the run should be retried an error was raised. Now, if a run is deleted while the re-execution system is determining whether the run should be retried, the run will not be retried.