[components] Inline the Component subclass discovery logic for Component discovery (#28120)
## Summary & Motivation

In 0.1.17 relies on behavior of modified utility
`find_subclasses_in_module` in `dagster`. But a `dagster` with the
updated utility is not published yet, leading to an error using newest
`dg` in the wild. This reverts the change to the utility function and
inlines the logic in `dagster-components`.

## How I Tested These Changes

Existing test suite + manually using editable `dagster-components` w/
published `dagster`.