Support `tags` for freshness checks sensor factory (#28131)
## Summary & Motivation

Interface for
[build_sensor_for_freshness_checks](https://github.com/dagster-io/dagster/blob/64ea57c11c086e91523525011eb52df006bb126d/python_modules/dagster/dagster/_core/definitions/asset_check_factories/freshness_checks/sensor.py#L38)
is really limited; this helps add some flexibility.

## How I Tested These Changes

I don't know if it's necessary to add a whole test for this. I could
modify one of the existing tests and check that the `run_request` in the
test has the right tags?

## Changelog

Added support for passing `tags` to the created `RunRequest` when using
`build_sensor_for_freshness_checks()`.