[components] Fix flaky module -> path resolution (#28127)
## Summary & Motivation

There was a method of resolution from modules to paths that relied on
using `importlib.util.find_spec_from_module`. This was behaving flakily
in certain test contexts, leading to mysterious failures. This PR
changes the method of resolution to skip importlib and instead just
perform a straightforward conversion to path. It only works for modules
that are part of the project root but that's the only thing we were
using it for anyway.

## How I Tested These Changes

Reenabled test that was failing before.