[dg] Implement scaffoldable as generic module to replace component-specific scaffolding features (#28162)
## Summary & Motivation

This stack changes our scaffolding infrastructure in components to a lightweight, composable framework that could be reused in other other contexts. 

We plan on extending the scaffolding framework to concepts like the project and possibly vanilla definitions.

Even if we do not do that in the short-term, I believe this is a superior structure that makes the Components base class more focused and leaves scaffodling as a standalone component.

The first step is to implement the replacement decorator `scaffoldable` and to move the generic scaffodling classes (to be renamed upstack) to `dagster_component.scaffoldable`. The eventual home of scaffoldable will be `dagster.scaffoldable` or possibly a standalone package.

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG