Add --use-editable-components-only flag (#28140)
## Summary & Motivation

Adds a `--use-editable-components-only` flag to complement
`--use-editable-dagster`.

- `--use-editable-components-only` is an option to commands that
scaffold projects `dg init` and `dg project scaffold`. It will bring in
`dagster_components` as an editable install but _not_ `dagster` and
other libs. This simulates the environment into which dg/components is
released, so long as it is on a separate release schedule from Dagster
core.
- Update all tests that were using `--use-editable-dagster` to instead
use `--use-editable-components-only`.
- Refactor these options into a shared group.

## How I Tested These Changes

New unit tests.