[ui] Replace references with data-testId => data-testid (#28159)
These are interchangeable in the tests (they both work), but at some
point React started printing errors about data-testId in the console:

```Warning: React does not recognize the `data-testId` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `data-testid` instead. If you accidentally passed it from a parent component, remove it from the DOM element.```

Co-authored-by: bengotow <bgotow@elementl.com>