[components] Switch dagster-dg tests to explicitly opt in to fixed test components (#28157)
## Summary & Motivation

Now that the test components live in `dagster-test` instead of
`dagster-components`, we don't want to target them by default in tests
since the `dagster-test` package is not installed in non-editable
scaffolded projects. This was leading to confusing error messages about
`dagster-test` not being installed in new tests.

## How I Tested These Changes

Existing test suite.