[components] Read project code location name/target from `tool.dagster` (#28171)
## Summary & Motivation

There was a discrepancy where the
`code_location_load_target_module_name` was hardcoded to
`<root_module>/definitions.py` instead of being read from
`tool.dagster`. This changes it to read from `tool.dagster`. This is an
intermediate step to the elimination of `tool.dagster` from projects.

## How I Tested These Changes

New unit test.