[dagster-fivetran] Handle missing schemas for unsynced connectors (#28080)
## Summary & Motivation

Fixes AD-881.

The `schemas` field is marked as required in Fivetran docs for the
`connections/{connection_id}/schemas`, but is missing for any connection
that has not already been synced.

Updating the logic to consider schemas as optional, and to discard
connectors without schemas/that have not been synced yet.

## How I Tested These Changes

Additional test with BK