[ui] Make asset overview show metadata, schema of latest event, not “latest partition or null" (#28174)
## Summary & Motivation

Fixes
https://linear.app/dagster-labs/issue/FE-589/metadata-plots-only-work-in-asset-catalog-when-most-recent-partition,
https://github.com/dagster-io/dagster/issues/24600


![image](https://github.com/user-attachments/assets/fe35cfca-dd56-4123-8c9e-2db2a12d2a4a)

It was flagged by a user that the new Asset Overview page shows the
metadata and schema for the most recent PARTITION, or empty states if
that partition is not present. This is somewhat confusing because these
sections can render empty while the tag at the top still shows the
latest event. (Screenshot above)

I think at some point the tag at the top of the page was also restricted
to showing the latest partition, so this made sense. However, it seems
problematic because for any daily partitioned asset, it is missing for
some portion of the day and all the information disappears for that
period.

The fix in this PR is just to make the metadata + schema portions of the
page use the latest event, even if it's not for a recent / today's
partition. This could have slightly undesirable behavior during
backfills (eg: showing you metadata for an older partition), but means
that the materialization in the green pill and the materialization used
for these sections is always the same.

## How I Tested These Changes

Tested with a manual repro

## Changelog

[ui] The asset overview tab for a partitioned asset now shows metadata
and schema of the most recent materialization, not today's partition.

---------

Co-authored-by: bengotow <bgotow@elementl.com>