[components] Ignore and warn about `tool.dg.cli` in project inside workspace (#28173)
## Summary & Motivation

We permit `tool.dg.cli` in standalone project configs, but when a
project is inside a workspace we want to use only the `tool.dg.cli`
section in the workspace config to avoid ambiguity about which section
should apply when operating on a project from within a workspace. This
PR emits a warning and ignores `tool.dg.cli` in a project config that is
inside a workspace.

## How I Tested These Changes

New unit test.