[dg] Demonstrate pattern for how we can support sequences generically (#28255)
## Summary & Motivation

This is not perfect yet, but provides a fairly condense syntax for dealing with collection types.

What we really need is a more dynamic engine that knows how to descend over the built-in `typing` types.

So that 

```python
    @dataclass
    class SomeObject(ResolvableFromSchema[SomeObjectSchema]):
        specs: Annotated[
            Sequence[AssetSpec],
            DSLFieldResolver(AssetSpecResolutionSpec.resolver_fn(AssetSpec).from_seq),
        ]
```

Can become

```python

ResolvedAssetSpec: TypeAlias = Annotated[AssetSpec, DSLFieldResolver(...)]     # exact syntax tbd

    @dataclass
    class SomeObject(ResolvableFromSchema[SomeObjectSchema]):
        specs:  Sequence[ResolvedAssetSpec]
```

but we can do that as a followup

## How I Tested These Changes

BK

## Changelog

NOCHANGELOG