[utils] Update pyproject.toml with exposed API lifecycle warnings (#28193)
## Summary & Motivation

Initially done in #28099. 

The warnings must be exposed in the latest version in pypi before
updating `pyproject.toml`, otherwise tests are failing for our cloud
templates, see
[here](https://buildkite.com/dagster/dagster-dagster/builds/113138#_) -
project.toml must be compatible with the latest version of Dagster in
pypi when users scaffold projects using these templates.

To be merged when #28099 is released.

## How I Tested These Changes

BK