[components] Clean up TOML (#28146)
## Summary & Motivation

Cleanup/enhancement of some of the TOML utilities we use for validation
and modification of TOML files. The enhancements are needed upstack when
we start adding `[[tool.dg.workspace.projects]]` entries.

- Change utility method `modify_pyproject.toml` to more general
`modify_toml`
- Add `modify_toml_as_dict` to get a plain dict instead of a TOML
document object. This is useful in config testing where it's fine to
nuke styling and write the modified dict back as new TOML. Modifying the
`TOMLDocument` object is more involved.
- Allow TOML paths used for TOML accessors to include array indices

This entire TOML util section will leave `dg` and become a part of a
factored out utility package when we do that (if we cannot find a
well-maintained library that supports this stuff, which I was unable to
find-- we need `tomlkit` for style/comment preservation when we modify
TOML files but it's pretty low level).

## How I Tested These Changes

Existing test suite.