[RFC] [components] Make pyproject.toml source of truth for deployment code locations (BUILD-652) (#27477)
## Summary & Motivation

- Make the `tool.dg.workspace.projects` list the source of truth for the
projects in a workspace. This replaces the behavior of just relying on
the content of the hardcoded "projects" folder of the workspace.
- The elements of `tool.dg.workspace.projects` are called "project
specs"-- currently there is just a single key, `path`, a path relative
to the workspace root.
- If you choose to have `dg init` scaffold a project, it will always do
so inside a created "projects" folder.
- `dg scaffold project` now expects its argument to be a path rather
than just a name. Intermediate folders will be created for you.
- The `dg list project` command now prints a list of project paths
instead of names.

## How I Tested These Changes

New unit tests.