[Asset Graph] Show an indeterminate loading bar instead of un-rendering graph and showing spinner. (#28336)
## Summary & Motivation

Currently when you change filters in the asset graph it immediately
shows a spinner while it calculates the new graph. One unintended side
effect of this is that in the filter menu if you check a filter we
immediately show the spinner and effectively only allow you to select 1
filter at a time. To remedy this, instead of showing the spinner, lets
just add an indeterminate loading bar and keep the previous graph
rendered until the new one is ready. This allows users to move fast
while filtering without waiting for potentially slow layout
re-rendering.

## How I Tested These Changes

locally


https://github.com/user-attachments/assets/39a9147a-6a43-403d-a683-1160b42fb9b3