[resources] support async yield_for_execution (#28142)
Add support for async context manager backed resources by passing the
same event loop to resource init and core execution.

The biggest bit of awkwardness here is that to ensure the same event
loop is used for both resource init and execution in the direct
execution case you have to manually pass the loop in to context creation
and then use it to execute your asset/op. This is demonstrated in the
test.

resolves https://github.com/dagster-io/dagster/issues/27974

## How I Tested These Changes

added tests

## Changelog

* `async def yield_for_execution` is now supported on
`ConfigurableResource`. An `event_loop` argument has been added to
context builders to support direct execution.