dg launch assets command (#28265)
Summary:
This was helpful for understanding the challenges for making a dg experience that will transition seamlessly between OSS and cloud.

The existing `dagster asset materialize` command is not workspace (or pyproject.toml) aware and requires passing in the pointer to a single code location. It does have the nice property that it synchronously executes the run and displays it in the CLI while it executes (compare with, say, the `dagster job launch` command, which submits a run but then returns right away with no logs before it executes).

Longer term to give an experience that gives you the best of both worlds in both OSS and cloud, we need an interface that is workspace-aware (in both OSS and cloud) and can handle launching a run (or backfill!) with the selected assets and configuration, and also optionally streaming the logs from that run or backfill over the CLI until it is finished. But this is a start as a first step.

## Summary & Motivation

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.