Fix issue with cycle resolution when using additional_deps (#28332)
## Summary & Motivation

The key issue here is that we were using `keys_by_input_name` as a way to determine which execution dependencies we cared about, when we actually have a kinda nuanced system for determining this, which depends on if the node contains only asset checks or not.

The code here is fairly verbose, but I wanted to factor it in a way that maximized explainability, given how weird this stuff is.

## How I Tested These Changes

## Changelog

Fixed a bug that could cause invalid circular dependency errors when using asset checks with additional dependencies.